refactor(app_user): resolve architecture audit cross-feature imports (#381)
- Move explore feature providers/logic to app-level shared locations
  (features/explore/ → logic/ and widgets/) to fix home→explore and
  search→explore cross-feature import violations
- Add GoRouter routes for PrivacyPage and BlockedPartnersPage to fix
  home→settings cross-feature import (Navigator.push → route-based nav)
- Extract skeleton classes from event_detail_content.dart (713→513 lines)
  into separate part file event_detail_content_skeleton.dart
- Update CLAUDE.md table count from 29 to 50 to match backend.md

Closes #381

Constraint: part-of files share scope — _SliverTabBarDelegate stays in main file since it's used by _EventDetailContent
Rejected: Move providers to home feature | search feature also depends on them
Rejected: Keep explore directory with barrel exports | doesn't fix structural violation
Confidence: high
Scope-risk: moderate
