+++ apps/app_user/lib/src/features/event/detail/event_detail_page.dart
@@ -15,6 +15,7 @@ import 'package:minglit_kit/minglit_kit.dart';
 
 part 'event_bottom_ticket_bar.dart';
 part 'event_detail_content.dart';
+part 'event_detail_content_skeleton.dart';
 part 'event_entry_conditions_section.dart';
 part 'event_info_tile.dart';
 part 'event_quill_viewer.dart';
