+++ apps/app_user/lib/src/features/search/search_page.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 
 import 'package:app_user/src/features/event/logic/event_coordinator.dart';
-import 'package:app_user/src/features/explore/providers/explore_state_provider.dart';
+import 'package:app_user/src/logic/feed_state_provider.dart';
 import 'package:flutter/material.dart';
 import 'package:minglit_kit/minglit_kit.dart';
