+++ apps/app_user/lib/src/logic/eligibility_filter.dart
