+++ apps/app_user/lib/src/widgets/explore_filter_chip_bar.dart
@@ -1,4 +1,4 @@
-import 'package:app_user/src/features/explore/providers/explore_state_provider.dart';
+import 'package:app_user/src/logic/feed_state_provider.dart';
 import 'package:flutter/material.dart';
 import 'package:minglit_kit/minglit_kit.dart';
