+++ apps/app_user/test/integration/flow_event_browse_test.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 
 import 'package:app_user/src/features/event/detail/event_detail_page.dart';
-import 'package:app_user/src/features/explore/providers/explore_state_provider.dart';
+import 'package:app_user/src/logic/feed_state_provider.dart';
 import 'package:app_user/src/features/partner/detail/partner_detail_page.dart';
 import 'package:app_user/src/features/partner/detail/partner_events_page.dart';
 import 'package:flutter_test/flutter_test.dart';
