+++ apps/app_user/test/src/features/explore/logic/eligibility_filter_test.dart
@@ -1,4 +1,4 @@
-import 'package:app_user/src/features/explore/logic/eligibility_filter.dart';
+import 'package:app_user/src/logic/eligibility_filter.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:minglit_kit/minglit_kit.dart';
