+++ apps/app_user/test/src/features/explore/providers/explore_filter_integration_test.dart
@@ -1,4 +1,4 @@
-import 'package:app_user/src/features/explore/providers/explore_state_provider.dart';
+import 'package:app_user/src/logic/feed_state_provider.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:minglit_kit/minglit_kit.dart';
