+++ apps/app_user/test/src/features/explore/providers/recommendation_feed_notifier_test.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 
-import 'package:app_user/src/features/explore/logic/eligibility_filter.dart';
-import 'package:app_user/src/features/explore/providers/explore_state_provider.dart';
+import 'package:app_user/src/logic/eligibility_filter.dart';
+import 'package:app_user/src/logic/feed_state_provider.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:minglit_kit/minglit_kit.dart';
 import 'package:mocktail/mocktail.dart';
